Update README.md
This is a complete c++ program which is a version of musical chairs, 
implemented using a circular linked list.  The program asks the user for the number of
players (2 to 7) and then for their names.  The program then selects a random number of
players to loop through (from a single player to three times the initial number of players).
It then displays a list of the players, with arrows pointing to the current player, 
and moves through the current players the chosen number of times. The player who is chosen
at this point will be eliminated.  This continues until there is one player remaining. 
At the end of each round, the program prints the name of the eliminated player.  
At the end of the game, the program prints the name of the winner.